+++ Raylib-cs/types/Ray.cs
@@ -4,52 +4,6 @@ using System.Runtime.InteropServices;
 namespace Raylib_cs
 {
     /// <summary>
-    /// Rectangle type
-    /// </summary>
-    [StructLayout(LayoutKind.Sequential)]
-    public struct Rectangle
-    {
-        public float x;
-        public float y;
-        public float width;
-        public float height;
-
-        public Rectangle(float x, float y, float width, float height)
-        {
-            this.x = x;
-            this.y = y;
-            this.width = width;
-            this.height = height;
-        }
-
-        public override string ToString()
-        {
-            return $"{{X:{x} Y:{y} Width:{width} Height:{height}}}";
-        }
-    }
-
-    /// <summary>Bounding box type</summary>
-    [StructLayout(LayoutKind.Sequential)]
-    public struct BoundingBox
-    {
-        /// <summary>
-        /// Minimum vertex box-corner
-        /// </summary>
-        public Vector3 min;
-
-        /// <summary>
-        /// Maximum vertex box-corner
-        /// </summary>
-        public Vector3 max;
-
-        public BoundingBox(Vector3 min, Vector3 max)
-        {
-            this.min = min;
-            this.max = max;
-        }
-    }
-
-    /// <summary>
     /// Ray, ray for raycasting
     /// </summary>
     [StructLayout(LayoutKind.Sequential)]
